rename component file in index.ts
